[refactor] Remove old schema change rollup backend decommission code (#8030)
