+++ fe/fe-core/src/main/java/org/apache/doris/common/proc/BackendsProcDir.java
@@ -17,7 +17,7 @@
 
 package org.apache.doris.common.proc;
 
-import org.apache.doris.alter.DecommissionBackendJob.DecommissionType;
+import org.apache.doris.alter.DecommissionType;
 import org.apache.doris.catalog.Catalog;
 import org.apache.doris.cluster.Cluster;
 import org.apache.doris.common.AnalysisException;
