+++ fe/fe-core/src/main/java/org/apache/doris/master/MasterImpl.java
@@ -17,12 +17,7 @@
 
 package org.apache.doris.master;
 
-import org.apache.doris.alter.AlterJob;
 import org.apache.doris.alter.AlterJobV2.JobType;
-import org.apache.doris.alter.MaterializedViewHandler;
-import org.apache.doris.alter.RollupJob;
-import org.apache.doris.alter.SchemaChangeHandler;
-import org.apache.doris.alter.SchemaChangeJob;
 import org.apache.doris.catalog.Catalog;
 import org.apache.doris.catalog.Database;
 import org.apache.doris.catalog.MaterializedIndex;
@@ -47,12 +42,10 @@ import org.apache.doris.task.CheckConsistencyTask;
 import org.apache.doris.task.ClearAlterTask;
 import org.apache.doris.task.CloneTask;
 import org.apache.doris.task.CreateReplicaTask;
-import org.apache.doris.task.CreateRollupTask;
 import org.apache.doris.task.DirMoveTask;
 import org.apache.doris.task.DownloadTask;
 import org.apache.doris.task.PublishVersionTask;
 import org.apache.doris.task.PushTask;
-import org.apache.doris.task.SchemaChangeTask;
 import org.apache.doris.task.SnapshotTask;
 import org.apache.doris.task.UpdateTabletMetaInfoTask;
 import org.apache.doris.task.UploadTask;
@@ -177,16 +170,8 @@ public class MasterImpl {
                     finishDropReplica(task);
                     break;
                 case SCHEMA_CHANGE:
-                    Preconditions.checkState(request.isSetReportVersion());
-                    checkHasTabletInfo(request);
-                    finishTabletInfos = request.getFinishTabletInfos();
-                    finishSchemaChange(task, finishTabletInfos, request.getReportVersion());
-                    break;
                 case ROLLUP:
-                    checkHasTabletInfo(request);
-                    finishTabletInfos = request.getFinishTabletInfos();
-                    finishRollup(task, finishTabletInfos);
-                    break;
+                    throw new RuntimeException("Schema change and rollup job is not used any more, use alter task instead");
                 case CLONE:
                     finishClone(task, request);
                     break;
@@ -458,10 +443,11 @@ public class MasterImpl {
         }
         MaterializedIndex index = partition.getIndex(indexId);
         if (index == null) { 
-            // this means the index is under rollup
-            MaterializedViewHandler materializedViewHandler = Catalog.getCurrentCatalog().getRollupHandler();
-            AlterJob alterJob = materializedViewHandler.getAlterJob(olapTable.getId());
-            if (alterJob == null && olapTable.getState() == OlapTableState.ROLLUP) {
+            // In alter job v2 case
+            // alter job is always == null, so that we could remove the condition
+            // if alter job is always null, then could not covert it to a rollup
+            // job, will throw exception, so just throw exception in this case
+            if (olapTable.getState() == OlapTableState.ROLLUP) {
                 // this happens when:
                 // a rollup job is finish and a delete job is the next first job (no load job before)
                 // and delete task is first send to base tablet, so it will return 2 tablets info.
@@ -472,14 +458,7 @@ public class MasterImpl {
                 LOG.warn("Cannot find table[{}].", olapTable.getId());
                 return null;
             }
-            RollupJob rollupJob = (RollupJob) alterJob;
-            MaterializedIndex rollupIndex = rollupJob.getRollupIndex(partition.getId());
-
-            if (rollupIndex == null) {
-                LOG.warn("could not find index for tablet {}", tabletId);
-                return null;
-            }
-            index = rollupIndex;
+            throw new MetaNotFoundException("Could not find related replica");
         }
         Tablet tablet = index.getTablet(tabletId);
         if (tablet == null) {
@@ -671,38 +650,13 @@ public class MasterImpl {
                 LOG.warn("tablet[{}] may be dropped. push index[{}]", tabletId, pushIndexId);
                 return null;
             }
-
-            MaterializedViewHandler materializedViewHandler = Catalog.getCurrentCatalog().getRollupHandler();
-            AlterJob alterJob = materializedViewHandler.getAlterJob(olapTable.getId());
-            if (alterJob == null) {
-                // this happens when:
-                // a rollup job is finish and a delete job is the next first job (no load job before)
-                // and delete task is first send to base tablet, so it will return 2 tablets info.
-                // the second tablet is rollup tablet and it is no longer exist in alterJobs queue.
-                // just ignore the rollup tablet info. it will be handled in rollup tablet delete task report.
-
-                // add log to observe
-                LOG.warn("Cannot find table[{}].", olapTable.getId());
-                return null;
-            }
-
-            ((RollupJob) alterJob).updateRollupReplicaInfo(partition.getId(), indexId, tabletId, backendId,
-                                                           schemaHash, version, versionHash, rowCount, dataSize);
-            // replica info is saved in rollup job, not in load job
             return null;
         }
 
         int currentSchemaHash = olapTable.getSchemaHashByIndexId(pushIndexId);
         if (schemaHash != currentSchemaHash) {
             if (pushState == PartitionState.SCHEMA_CHANGE) {
-                SchemaChangeHandler schemaChangeHandler = Catalog.getCurrentCatalog().getSchemaChangeHandler();
-                AlterJob alterJob = schemaChangeHandler.getAlterJob(olapTable.getId());
-                if (alterJob != null && schemaHash != ((SchemaChangeJob) alterJob).getSchemaHashByIndexId(pushIndexId)) {
-                    // this is a invalid tablet.
-                    throw new MetaNotFoundException("tablet[" + tabletId
-                            + "] schemaHash is not equal to index's switchSchemaHash. "
-                            + ((SchemaChangeJob) alterJob).getSchemaHashByIndexId(pushIndexId) + " vs. " + schemaHash);
-                }
+                // Alter job is always null, so that not deal with it
             } else {
                 // this should not happen. observe(cmy)
                 throw new MetaNotFoundException("Diff tablet[" + tabletId + "] schemaHash. index[" + pushIndexId + "]: "
@@ -734,28 +688,6 @@ public class MasterImpl {
 
     private void finishDropReplica(AgentTask task) {
         AgentTaskQueue.removeTask(task.getBackendId(), TTaskType.DROP, task.getSignature());
-    }
-
-    private void finishSchemaChange(AgentTask task, List<TTabletInfo> finishTabletInfos, long reportVersion)
-            throws MetaNotFoundException {
-        Preconditions.checkArgument(finishTabletInfos != null && !finishTabletInfos.isEmpty());
-        Preconditions.checkArgument(finishTabletInfos.size() == 1);
-
-        SchemaChangeTask schemaChangeTask = (SchemaChangeTask) task;
-        SchemaChangeHandler schemaChangeHandler = Catalog.getCurrentCatalog().getSchemaChangeHandler();
-        schemaChangeHandler.handleFinishedReplica(schemaChangeTask, finishTabletInfos.get(0), reportVersion);
-        AgentTaskQueue.removeTask(task.getBackendId(), TTaskType.SCHEMA_CHANGE, task.getSignature());
-    }
-
-    private void finishRollup(AgentTask task, List<TTabletInfo> finishTabletInfos)
-            throws MetaNotFoundException {
-        Preconditions.checkArgument(finishTabletInfos != null && !finishTabletInfos.isEmpty());
-        Preconditions.checkArgument(finishTabletInfos.size() == 1);
-
-        CreateRollupTask createRollupTask = (CreateRollupTask) task;
-        MaterializedViewHandler materializedViewHandler = Catalog.getCurrentCatalog().getRollupHandler();
-        materializedViewHandler.handleFinishedReplica(createRollupTask, finishTabletInfos.get(0), -1L);
-        AgentTaskQueue.removeTask(task.getBackendId(), TTaskType.ROLLUP, task.getSignature());
     }
 
     private void finishClone(AgentTask task, TFinishTaskRequest request) {
