+++ fe/fe-core/src/main/java/org/apache/doris/system/Backend.java
@@ -17,7 +17,7 @@
 
 package org.apache.doris.system;
 
-import org.apache.doris.alter.DecommissionBackendJob.DecommissionType;
+import org.apache.doris.alter.DecommissionType;
 import org.apache.doris.catalog.Catalog;
 import org.apache.doris.catalog.DiskInfo;
 import org.apache.doris.catalog.DiskInfo.DiskState;
