+++ fe/fe-core/src/test/java/org/apache/doris/catalog/FakeEditLog.java
@@ -19,8 +19,6 @@ package org.apache.doris.catalog;
 
 import org.apache.doris.alter.AlterJobV2;
 import org.apache.doris.alter.BatchAlterJobPersistInfo;
-import org.apache.doris.alter.RollupJob;
-import org.apache.doris.alter.SchemaChangeJob;
 import org.apache.doris.cluster.Cluster;
 import org.apache.doris.persist.BatchRemoveTransactionsOperation;
 import org.apache.doris.persist.EditLog;
@@ -67,28 +65,6 @@ public class FakeEditLog extends MockUp<EditLog> {
     @Mock
     public void logCreateCluster(Cluster cluster) {
         // do nothing
-    }
-    
-    @Mock
-    public void logStartRollup(RollupJob rollupJob) {
-        
-    }
-
-    @Mock
-    public void logFinishingRollup(RollupJob rollupJob) {
-        
-    }
-    
-    @Mock
-    public void logCancelRollup(RollupJob rollupJob) {
-    }
-    
-    @Mock
-    public void logStartSchemaChange(SchemaChangeJob schemaChangeJob) {
-    }
-    
-    @Mock
-    public void logFinishingSchemaChange(SchemaChangeJob schemaChangeJob) {
     }
     
     @Mock
